+++ include/md/forcefield/bonded_pairwise_forcefield.hpp
@@ -1,4 +1,4 @@
-// Copyright snsinfu 2019.
+// Copyright snsinfu 2019-2021.
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
@@ -112,9 +112,9 @@ namespace md
         {
         }
 
-        auto bonded_pairwise_potential(md::system const&, md::index i, md::index j) const
+        auto bonded_pairwise_potential(md::system const& system, md::index i, md::index j) const
         {
-            return potfun_(i, j);
+            return potfun_(system, i, j);
         }
 
     private:
@@ -126,7 +126,7 @@ namespace md
     template<typename P>
     auto make_bonded_pairwise_forcefield(P pot)
     {
-        auto potfun = detail::make_pair_potfun(pot);
+        auto potfun = detail::make_pair_potential_factory(pot);
         using potfun_type = decltype(potfun);
         return md::basic_bonded_pairwise_forcefield<potfun_type>{potfun};
     }