+++ include/md/forcefield/neighbor_pairwise_forcefield.hpp
@@ -1,4 +1,4 @@
-// Copyright snsinfu 2019.
+// Copyright snsinfu 2019-2021.
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
@@ -183,9 +183,9 @@ namespace md
         {
         }
 
-        auto neighbor_pairwise_potential(md::system const&, md::index i, md::index j) const
+        auto neighbor_pairwise_potential(md::system const& system, md::index i, md::index j) const
         {
-            return potfun_(i, j);
+            return potfun_(system, i, j);
         }
 
     private:
@@ -198,7 +198,7 @@ namespace md
     template<typename Box = md::open_box, typename P>
     auto make_neighbor_pairwise_forcefield(P pot)
     {
-        auto potfun = detail::make_pair_potfun(pot);
+        auto potfun = detail::make_pair_potential_factory(pot);
         using potfun_type = decltype(potfun);
         return md::basic_neighbor_pairwise_forcefield_impl<potfun_type, Box>{potfun};
     }