+++ tests/forcefield/test_bruteforce_pairwise_forcefield.cc
@@ -134,3 +134,29 @@ TEST_CASE("make_bruteforce_pairwise_forcefield - creates an bruteforce_pairwise_
     CHECK(std::is_same<pot_type, md::harmonic_potential>::value);
     CHECK(pot.spring_constant == 1.23);
 }
+
+TEST_CASE("make_bruteforce_pairwise_forcefield - accepts lambda(system, i, j)")
+{
+    auto forcefield = md::make_bruteforce_pairwise_forcefield(
+        [](md::system const&, md::index, md::index) {
+            return md::harmonic_potential{42};
+        }
+    );
+
+    md::system system;
+    md::harmonic_potential potential = forcefield.bruteforce_pairwise_potential(system, 0, 1);
+    CHECK(potential.spring_constant == 42);
+}
+
+TEST_CASE("make_bruteforce_pairwise_forcefield - accepts lambda(i, j)")
+{
+    auto forcefield = md::make_bruteforce_pairwise_forcefield(
+        [](md::index, md::index) {
+            return md::harmonic_potential{42};
+        }
+    );
+
+    md::system system;
+    md::harmonic_potential potential = forcefield.bruteforce_pairwise_potential(system, 0, 1);
+    CHECK(potential.spring_constant == 42);
+}